adds confirmation

+++ src/components/Festivals/FestivalItem.tsx
@@ -1,4 +1,4 @@
-import { BaseSyntheticEvent } from "react"
+import { BaseSyntheticEvent, useState } from "react"
 import { Link } from "react-router-dom"
 import * as api from '../../utils/api'
 
@@ -10,17 +10,26 @@ export interface Festival {
 }
 
 export function FestivalItem ({ festival, setFestivals }: {festival: Festival, setFestivals: Function}) {
+  const [confirmation, setConfirmation] = useState(false);
 
   const handleDelete = async (event: BaseSyntheticEvent) => {
     event.preventDefault();
-    await api.deleteFestivalByName(festival.festival_name);
-    setFestivals((prev: Festival[]) => {
-      const newFestivals = [...prev].filter((festivalToCheck: Festival) => festival.festival_key !== festivalToCheck.festival_key);
-      return newFestivals;
-    })
+    if (confirmation) {
+      await api.deleteFestivalByName(festival.festival_name);
+      setFestivals((prev: Festival[]) => {
+        const newFestivals = [...prev].filter((festivalToCheck: Festival) => festival.festival_key !== festivalToCheck.festival_key);
+        return newFestivals;
+      })
+      return;
+    }
+    setConfirmation(true);
   }
 
   return (
-    <li key={festival.festival_key}><Link to={`/festival_specs_fe/festivals/${festival.festival_name}/stages`}>{festival.festival_name} {new Date (festival.start_date).toDateString()} {new Date (festival.end_date).toDateString()}</Link><button onClick={handleDelete}>Delete</button></li>
+    <li key={festival.festival_key}>
+      <Link to={`/festival_specs_fe/festivals/${festival.festival_name}/stages`}>{festival.festival_name} {new Date (festival.start_date).toDateString()} {new Date (festival.end_date).toDateString()}</Link>
+      <button onClick={handleDelete}>{!confirmation ? 'Delete' : 'Confirm'}</button>
+      {confirmation && <button onClick={() => setConfirmation(false)}>Reset</button>}
+    </li>
   )
 }
